cpu_worker/__init__.py
#!/usr/bin/env python3
"""
bildwerk CPU Worker - OpenVINO-based image processing
"""

import asyncio
import aiohttp
from aiohttp import web
import yaml
import json
import os
import logging
from datetime import datetime
from pathlib import Path
from prometheus_client import start_http_server, Counter, Gauge, Histogram

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger('bildwerk-cpu-worker')

# Prometheus metrics
CPU_JOBS_PROCESSED = Counter('bildwerk_cpu_jobs_processed', 'Total CPU jobs processed')
CPU_JOBS_FAILED = Counter('bildwerk_cpu_jobs_failed', 'Total CPU jobs failed')
CPU_JOBS_IN_PROGRESS = Gauge('bildwerk_cpu_jobs_in_progress', 'CPU jobs in progress')
CPU_JOB_DURATION = Histogram('bildwerk_cpu_job_duration_seconds', 'CPU job duration')

class CPUWorker:
    def __init__(self, config_path: str):
        with open(config_path, 'r') as f:
            self.config = yaml.safe_load(f)
        self.jobs = {}
        self.running = True
        
    async def handle_job_request(self, request):
        """Handle incoming job request"""
        job_data = await request.json()
        job_id = job_data.get('job_id')
        
        logger.info(f"Received job: {job_id}")
        CPU_JOBS_IN_PROGRESS.inc()
        
        try:
            # TODO: Implement actual OpenVINO processing
            # For now, simulate processing
            await asyncio.sleep(2)  # Simulate processing time
            
            result = {
                'job_id': job_id,
                'status': 'completed',
                'output_path': f"/tmp/bildwerk/output/{job_id}.png",
                'started_at': datetime.now().isoformat(),
                'completed_at': datetime.now().isoformat(),
                'metrics': {
                    'processing_time_ms': 2000
                }
            }
            
            CPU_JOBS_PROCESSED.inc()
            CPU_JOB_DURATION.observe(2)
            
            return web.json_response(result, status=200)
            
        except Exception as e:
            logger.error(f"Job {job_id} failed: {e}")
            CPU_JOBS_FAILED.inc()
            return web.json_response({'error': str(e)}, status=500)
        finally:
            CPU_JOBS_IN_PROGRESS.dec()
            
    async def health_check(self, request):
        """Health check endpoint"""
        return web.json_response({'status': 'healthy', 'type': 'cpu'})
        
    async def system_stats(self, request):
        """System stats endpoint"""
        return web.json_response({
            'type': 'cpu',
            'status': 'ready',
            'models_available': []
        })
        
    async def start(self):
        """Start the worker"""
        app = web.Application()
        app.router.add_post('/api/v1/jobs', self.handle_job_request)
        app.router.add_get('/health', self.health_check)
        app.router.add_get('/system_stats', self.system_stats)
        app.router.add_get('/metrics', self.prometheus_handler)
        
        # Start Prometheus on separate port
        prom_port = self.config.get('prometheus', {}).get('port', 8082)
        start_http_server(prom_port)
        logger.info(f"Prometheus metrics on port {prom_port}")
        
        # Start API server on configured port
        api_port = self.config.get('worker', {}).get('port', 8081)
        runner = web.AppRunner(app)
        await runner.setup()
        site = web.TCPSite(runner, '0.0.0.0', api_port)
        await site.start()
        logger.info(f"CPU Worker started on port {api_port}")
        
    async def prometheus_handler(self, request):
        """Prometheus metrics endpoint"""
        from prometheus_client import generate_latest
        return web.Response(body=generate_latest(), content_type='text/plain')

async def main():
    config_path = os.environ.get('BILDWORK_CONFIG', '/opt/bildwerk/cpu/config.yaml')
    
    if not os.path.exists(config_path):
        logger.error(f"Config not found: {config_path}")
        return
        
    worker = CPUWorker(config_path)
    await worker.start()
    
    # Keep running
    while worker.running:
        await asyncio.sleep(3600)

if __name__ == '__main__':
    asyncio.run(main())